W8-O-5d: Der Rechenkern rechnet in double - Stundenreihen, Akkumulatoren und die Summenfunktionen des BHKW-Plan-Ports
Anwenderentscheid vom 07.09.2026: "alles in double, ist kein Nachteil und
systematisch. Summenfunktionen aus Original BHKW-Plan ebenfalls double."
Damit ist die Empfehlung Q6 des Einheitenkonzepts ("die 59 Stundenreihen bleiben
float") ueberholt.

WAS UMGESTELLT IST
- Alle Zeitreihen des Laufs (8760 / 35040 / 365 / 168 / 12) und alle
  Akkumulatoren in EPOS.Kern/Allgemein/Simulation/, dazu Ganglinie,
  ProfilBedarf, ErdreichTemperatur, WaermequelleClass, Kanalsatz,
  SimulationRunner, ZeitreihenExtraktor und CsvExportClass (CsvSpalte).
- BhkwPlan.cs vollstaendig: VectorInit/WattToKw/VectorenAddieren/VectorSumme/
  Normieren/NetzverlusteC/MonatsSumme/Heapsort/StromWocheToJahr/StdWerte und die
  drei Physikfunktionen. Die bewusste Nachbildung des FPU-Verhaltens der alten
  BHKWPLAN.DLL (Zwischenwert in double, Ergebnis auf float zurueckgeschrieben)
  ist aufgegeben; Zwischenwert und Speicherzelle haben jetzt dieselbe Breite.
- Die Controller, die Reihen liefern oder verbrauchen (GanglinienAuswertung,
  Quellprofil, StromspeicherSim, GebaeudeBedarf, SimulationErgebnis,
  Klimaregion, BedarfsVorschau), EPOS.UI (QuelleErdreichDaten/-Dialog) und die
  WinForms-Huellen.
- Datenbank- und Dateigrenze: Convert.ToSingle/(float) beim Lesen aus
  IDatenzugriff/RecordSet/DataRow wird Convert.ToDouble - SQLite REAL ist
  double, der Lesepfad verliert keine Stellen mehr.
- SpeicherEngine.RasterAdapter: ZuViertelstundenDouble nimmt double[],
  ZuFloat/ZuDouble entfallen (sie waren die float-Grenze zum Hauptprojekt);
  an ihre Stelle tritt Kopie(double[]). Referenzlauf/Ergebnisexport: die
  Ueberladung Vektor(float[]) ist tot und faellt.

WAS float BLEIBT (drei belegte Grenzen)
- ChartRenderer: Bildpunkte, Strichstaerken, Schriftgroessen - SkiaSharp
  rechnet in float. Die DATENREIHEN, die der Renderer annimmt, waren schon
  immer double und bleiben es. ChartProben: 44 Bilder, alle gruen.
- SemantikIndex/SemantikModell: Einbettungsvektoren des KI-Wissens, kein
  Rechenweg.
- Typpruefungen auf boxed Datenbankwerte (v is float, typeof(float)) in
  ProjektExportImportCtrl, KomponentenUebernahmeCtrl, MerkmalUebernahmeCtrl,
  ParameterUebersichtCtrl, AnlagenEindeutigkeit, DublettenPruefung,
  PufferSpStammCtrl, Ergebnisexport - Absicherung, kein Rechenweg.
Im Kern faellt die Zahl der float-Fundstellen damit von 788 in 42 Dateien auf
148 in 15 Dateien (davon 120 im ChartRenderer, 12 in der KI, 11 Typpruefungen
und 5 Kommentare, die die Geschichte erzaehlen).

WAECHTER
Neu: EPOS.Kern.Tests/DoubleWacheTests - in EPOS.Kern/Allgemein/Simulation/ und
BhkwPlan.cs steht kein float mehr (Typname, Convert.ToSingle, MathF, f-Suffix),
mit drei Gegenproben und LEERER Ausnahmeliste. EinheitenWacheTests bleibt gruen.

ZWEI EINGEFRORENE ERWARTUNGEN NACHGEZOGEN
BedarfVerwaltungTests: die Brauchwasser-Jahressumme steht jetzt auf 742,9000
statt 742,9008 kWh. Die 0,0008 kWh waren die Rundung der 8760 float-Zellen; der
neue Wert trifft die Katalogmenge (0,7429 MWh x 1000) exakt - die zweite
Zusicherung derselben Tests, die bisher nur auf zwei Stellen passte.

+++ EPOS.Kern.Tests/BhkwLeistungsgrenzeTests.cs
@@ -55,17 +55,17 @@ namespace EPOS.Kern.Tests
         private const double EIGENER_WERT_PROZENT = 15.0;
 
         /// <summary>Elektrische Leistung der Zeile <see cref="MODUL_OHNE_EIGENEN_WERT"/> [kW].</summary>
-        private const float PEL_MODUL = 10f;
+        private const double PEL_MODUL = 10.0;
 
         /// <summary>Thermische Leistung derselben Zeile [kW].</summary>
-        private const float PTHERM_MODUL = 19f;
+        private const double PTHERM_MODUL = 19.0;
 
         /// <summary>
         /// Strombedarf der Probestunde [kWh] — <b>kleiner als 30 % von
         /// <see cref="PEL_MODUL"/></b> (2 &lt; 3) und damit genau der Bereich, in dem
         /// sich Grenze 0 und Grenze 30 % unterscheiden.
         /// </summary>
-        private const float KLEINER_STROMBEDARF = 2f;
+        private const double KLEINER_STROMBEDARF = 2.0;
 
         // =================================================================================
         // 1 — Die zwei Ebenen im Rechenweg (kein stiller Fallback mehr)
@@ -80,7 +80,7 @@ namespace EPOS.Kern.Tests
         {
             if (!_db.Vorhanden) return;
 
-            Assert.Equal(0f, Grenzfaktor(MODUL_OHNE_EIGENEN_WERT, projektwertProzent: 0));
+            Assert.Equal(0.0, Grenzfaktor(MODUL_OHNE_EIGENEN_WERT, projektwertProzent: 0));
         }
 
         /// <summary>
@@ -92,7 +92,7 @@ namespace EPOS.Kern.Tests
         {
             if (!_db.Vorhanden) return;
 
-            Assert.Equal(0.3f, Grenzfaktor(MODUL_OHNE_EIGENEN_WERT, projektwertProzent: 30), 5);
+            Assert.Equal(0.3, Grenzfaktor(MODUL_OHNE_EIGENEN_WERT, projektwertProzent: 30), 5);
         }
 
         /// <summary>
@@ -105,7 +105,7 @@ namespace EPOS.Kern.Tests
         {
             if (!_db.Vorhanden) return;
 
-            Assert.Equal(0.1f, Grenzfaktor(MODUL_OHNE_EIGENEN_WERT, projektwertProzent: 10), 5);
+            Assert.Equal(0.1, Grenzfaktor(MODUL_OHNE_EIGENEN_WERT, projektwertProzent: 10), 5);
         }
 
         /// <summary>
@@ -117,7 +117,7 @@ namespace EPOS.Kern.Tests
         {
             if (!_db.Vorhanden) return;
 
-            float erwartet = (float)(EIGENER_WERT_PROZENT / 100.0);
+            double erwartet = (double)(EIGENER_WERT_PROZENT / 100.0);
 
             Assert.Equal(erwartet, Grenzfaktor(MODUL_MIT_EIGENEM_WERT, projektwertProzent: 30), 5);
             Assert.Equal(erwartet, Grenzfaktor(MODUL_MIT_EIGENEM_WERT, projektwertProzent: 0), 5);
@@ -144,7 +144,7 @@ namespace EPOS.Kern.Tests
 
             // Mit 30 % bleibt der Motor stehen: Er kann 2 kW nicht liefern, ohne unter
             // seine Untergrenze von 3 kW zu gehen.
-            Assert.Equal(0f, mitGrenze30.stromproduktion[0]);
+            Assert.Equal(0.0, mitGrenze30.stromproduktion[0]);
 
             // Ohne Untergrenze deckt er den Bedarf genau.
             Assert.Equal(KLEINER_STROMBEDARF, ohneGrenze.stromproduktion[0], 3);
@@ -172,7 +172,7 @@ namespace EPOS.Kern.Tests
             sim.modeBHKW = betriebsart;
             sim.strombedarf[0] = KLEINER_STROMBEDARF;
 
-            Assert.Equal(0f, sim.bhkwGrenzL[0]);
+            Assert.Equal(0.0, sim.bhkwGrenzL[0]);
 
             double[] rest = new double[Kanal.ANZAHL];
             for (int k = 0; k < rest.Length; k++) rest[k] = 50.0;
@@ -181,12 +181,12 @@ namespace EPOS.Kern.Tests
             sim.Stunde_Bedarf(0, false, rest);
             sim.Stunde_Ende(0, 0.0);
 
-            Assert.True(float.IsFinite(sim.waermeproduktion[0]),
+            Assert.True(double.IsFinite(sim.waermeproduktion[0]),
                         "Waermeproduktion ist keine endliche Zahl.");
-            Assert.True(float.IsFinite(sim.stromproduktion[0]),
+            Assert.True(double.IsFinite(sim.stromproduktion[0]),
                         "Stromproduktion ist keine endliche Zahl.");
-            Assert.True(sim.waermeproduktion[0] >= 0f);
-            Assert.True(sim.stromproduktion[0] >= 0f);
+            Assert.True(sim.waermeproduktion[0] >= 0.0);
+            Assert.True(sim.stromproduktion[0] >= 0.0);
 
             foreach (double wert in rest)
                 Assert.True(double.IsFinite(wert), "Ein Restbedarf ist keine endliche Zahl.");
@@ -289,7 +289,7 @@ namespace EPOS.Kern.Tests
         /// (<c>Vorbereiten_Zweikanalig</c>) und gibt den aufgelösten Faktor des ersten
         /// Moduls zurück.
         /// </summary>
-        private static float Grenzfaktor(int idModul, int projektwertProzent)
+        private static double Grenzfaktor(int idModul, int projektwertProzent)
         {
             return Vorbereitet(idModul, projektwertProzent).bhkwGrenzL[0];
         }
